conftest.py: move utils to conftest.py
